Feedback based changes: 1. For consistency, changed /stream to /chat 2. Made request body consistent with OpenAI API 3. Updated helper text to educate the user and link to OpenAI docs 4.  Updated base url to be fetched based on current url. 5. Added check marks to copy buttons
